01_RAG_FAQ_Chatbot/requirements.txt
langchain
openai
huggingface_hub
sentence-transformers
faiss-cpu
python-dotenv
pandas 